test: add tests for the state consumers
